Update code for Addresses

+++ models/address.py
@@ -9,7 +9,7 @@ class Address(Base, BaseDBModel):
 
     id = Column(BigInteger, primary_key=True, index=True, unique=True)
     user_id = Column(BigInteger, nullable=False)
-    city = Column(Integer, nullable=False)
-    district = Column(Integer, nullable=False)
-    ward = Column(Integer, nullable=False)
+    city = Column(BigInteger, nullable=False)
+    district = Column(BigInteger, nullable=False)
+    ward = Column(BigInteger, nullable=False)
     detail = Column(String(255), nullable=False)
